+++ CHANGES.rst
@@ -4,5 +4,5 @@ Changelog
 1.0.0 (????-??-??)
 ------------------
 
-- Initial release: label-dist-ic, label-dist-is, label-dist-od
+- Initial release: label-dist-ic, label-dist-is, label-dist-od, to-annotation-overlay-od